feat : 회원 정보 관리 인수테스트 작성
- JWT 토큰 활용하여 회원정보 CRUD시 인증절차 거치도록 구현
- 인증이 필요한 컨트롤러들의 패턴을 BearerAuthInterceptor에 추가
- 인증정보(LoginMember)가 없을 경우 200 상태로 빈 멤버를 응답
- 인증정보가 틀릴 경우 401 응답
- 사용하지 않는 다른 인증 및 관련 테스트 삭제
- 서버 프로젝트 스코프 컨벤션 리팩토링 진행

+++ src/main/java/wooteco/subway/web/member/MemberController.java
@@ -1,45 +1,69 @@
 package wooteco.subway.web.member;
 
+import java.net.URI;
+
+import org.springframework.data.repository.query.Param;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
 import wooteco.subway.domain.member.Member;
 import wooteco.subway.service.member.MemberService;
 import wooteco.subway.service.member.dto.MemberRequest;
 import wooteco.subway.service.member.dto.MemberResponse;
 import wooteco.subway.service.member.dto.UpdateMemberRequest;
 
-import java.net.URI;
-
 @RestController
 public class MemberController {
-    private MemberService memberService;
+    private final MemberService memberService;
 
     public MemberController(MemberService memberService) {
         this.memberService = memberService;
     }
 
     @PostMapping("/members")
-    public ResponseEntity createMember(@RequestBody MemberRequest view) {
+    public ResponseEntity<Void> createMember(@RequestBody MemberRequest view) {
         Member member = memberService.createMember(view.toMember());
         return ResponseEntity
-                .created(URI.create("/members/" + member.getId()))
-                .build();
+            .created(URI.create("/members/" + member.getId()))
+            .build();
     }
 
     @GetMapping("/members")
-    public ResponseEntity<MemberResponse> getMemberByEmail(@RequestParam String email) {
-        Member member = memberService.findMemberByEmail(email);
+    public ResponseEntity<MemberResponse> getMemberByEmail(
+        @LoginMember Member member,
+        @Param("email") String email
+    ) {
+        if (!member.isAuthenticated(email)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         return ResponseEntity.ok().body(MemberResponse.of(member));
     }
 
     @PutMapping("/members/{id}")
-    public ResponseEntity<MemberResponse> updateMember(@PathVariable Long id, @RequestBody UpdateMemberRequest param) {
+    public ResponseEntity<MemberResponse> updateMember(
+        @LoginMember Member member,
+        @PathVariable Long id,
+        @RequestBody UpdateMemberRequest param
+    ) {
+        if (!member.isAuthenticated(id)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         memberService.updateMember(id, param);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/members/{id}")
-    public ResponseEntity<MemberResponse> deleteMember(@PathVariable Long id) {
+    public ResponseEntity<MemberResponse> deleteMember(@LoginMember Member member, @PathVariable Long id) {
+        if (!member.isAuthenticated(id)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         memberService.deleteMember(id);
         return ResponseEntity.noContent().build();
     }
